# docs/api/openapi-1.2.0.json
{
 "components": {
  "schemas": {
...
   "AlertHistory": {
    "additionalProperties": true,
    "properties": {
     "history": {
      "items": {
       "$ref": "#/components/schemas/AlertRow"
      },
      "title": "History",
      "type": "array"
     }
    },
    "required": [
     "history"
    ],
    "title": "AlertHistory",
    "type": "object"
   },
   "AlertRow": {
    "additionalProperties": true,
    "properties": {
     "channel": {
      "title": "Channel",
      "type": "string"
     },
     "detail": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Detail"
     },
     "event": {
      "title": "Event",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "status": {
      "title": "Status",
      "type": "string"
     },
     "ts": {
      "title": "Ts",
      "type": "number"
     }
    },
    "required": [
     "id",
     "ts",
     "channel",
     "event",
     "status"
    ],
    "title": "AlertRow",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "MaintenanceRequest": {
    "properties": {
     "enabled": {
      "title": "Enabled",
      "type": "boolean"
     }
    },
    "required": [
     "enabled"
    ],
    "title": "MaintenanceRequest",
    "type": "object"
   },
...
   "ServerAddRequest": {
    "properties": {
     "color": {
      "anyOf": [
       {
        "maxLength": 16,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Color"
     },
     "directory": {
      "maxLength": 400,
      "minLength": 1,
      "title": "Directory",
      "type": "string"
     },
     "id": {
      "anyOf": [
       {
        "maxLength": 64,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Id"
     },
     "jar": {
      "default": "",
      "maxLength": 180,
      "title": "Jar",
      "type": "string"
     },
     "name": {
      "maxLength": 60,
      "minLength": 1,
      "title": "Name",
      "type": "string"
     },
     "type": {
      "anyOf": [
       {
        "maxLength": 32,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Type"
     }
    },
    "required": [
     "name",
     "directory"
    ],
    "title": "ServerAddRequest",
    "type": "object"
   },
...
   "ServerList": {
    "additionalProperties": true,
    "properties": {
     "default": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Default"
     },
     "servers": {
      "items": {
       "$ref": "#/components/schemas/ServerRow"
      },
      "title": "Servers",
      "type": "array"
     }
    },
    "required": [
     "servers"
    ],
    "title": "ServerList",
    "type": "object"
   },
   "ServerRow": {
    "additionalProperties": true,
    "properties": {
     "color": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Color"
     },
     "default": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Default"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "max_players": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Players"
     },
     "minecraft_version": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Minecraft Version"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "players_online": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "description": "null until it has been measured",
      "title": "Players Online"
     },
     "players_verified": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Players Verified"
     },
     "state": {
      "title": "State",
      "type": "string"
     },
     "state_verified": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "State Verified"
     },
     "type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Type"
     },
     "type_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Type Name"
     },
     "uptime": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Uptime"
     }
    },
    "required": [
     "id",
     "name",
     "state"
    ],
    "title": "ServerRow",
    "type": "object"
   },
...
   "SettingsRequest": {
    "properties": {
     "updates": {
      "additionalProperties": true,
      "title": "Updates",
      "type": "object"
     }
    },
    "required": [
     "updates"
    ],
    "title": "SettingsRequest",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "Minecraft Server Control",
  "version": "1.2.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/maintenance": {
   "post": {
    "operationId": "set_maintenance_api_maintenance_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/MaintenanceRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Set Maintenance"
   }
  },
...
  "/api/notifications/history": {
   "get": {
    "operationId": "notification_history_api_notifications_history_get",
    "parameters": [
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 50,
       "title": "Limit",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AlertHistory"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Notification History"
   }
  },
  "/api/notifications/test": {
   "post": {
    "operationId": "test_notification_api_notifications_test_post",
    "parameters": [
     {
      "in": "query",
      "name": "channel",
      "required": true,
      "schema": {
       "title": "Channel",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Test Notification"
   }
  },
...
  "/api/ports/suggest": {
   "get": {
    "operationId": "suggest_port_api_ports_suggest_get",
    "parameters": [
     {
      "in": "query",
      "name": "protocol",
      "required": false,
      "schema": {
       "default": "tcp",
       "title": "Protocol",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Suggest Port"
   }
  },
...
  "/api/servers": {
   "get": {
    "description": "Every registered server with its measured state, players and uptime.",
    "operationId": "servers_api_servers_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ServerList"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "Servers"
   },
   "post": {
    "description": "Register an existing Minecraft server folder. Nothing in the folder is\nchanged and Minecraft is not started.",
    "operationId": "add_server_api_servers_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ServerAddRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Add Server"
   }
  },
  "/api/servers/{server_id}": {
   "delete": {
    "description": "Take a server off the list. Its folder, world and backups stay.",
    "operationId": "remove_server_api_servers__server_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "server_id",
      "required": true,
      "schema": {
       "title": "Server Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Remove Server"
   }
  },
...
  "/api/settings": {
   "get": {
    "operationId": "get_settings_api_settings_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "Get Settings"
   },
   "put": {
    "operationId": "update_settings_api_settings_put",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SettingsRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Update Settings"
   }
  },
...
 }
}